app/requirements.txt
websockets
python-dotenv
pyaudio
numpy
sounddevice